fix: pay a trade total wider than a slot without narrowing it to a uint16
The buying branch compared the purse against a uint32 total and then spent
uint16(total), so a purchase over 65,535 silver was settled for the remainder
of the division. heldInPack sums silver across pack slots and reports a uint32
for that reason; the spending rule now takes one too, so the comparison and the
spend are the same width and no narrowing is left on the paying side.

The comment that stood there argued the opposite of what is true — forty slots
of a uint16 count do not fit a uint16, which is why the truncation was
reachable rather than why it was not.

A restored pack is what makes it reachable: restoredSlots copies a stored count
straight from the record and validateStoredSlot bounds it by nothing but the
uint16 it is stored in, so a purse is not held to a stack maximum.

The selling branch keeps its explicit MaxUint16 refusal: insert writes a single
slot's count and genuinely takes a uint16, which is a different bound from what
a pack-wide spend needs.

+++ server/internal/game/inventory.go
@@ -275,7 +275,7 @@ func (t *slotTable) insert(itemID ItemID, count uint16) uint16 {
 // distinction rather than a detail.** Nothing crafted has ever been worn, so "wherever it
 // is" and "wherever the insertion rule could have put it" were the same set of slots for
 // as long as this had one caller.
-func (t *slotTable) consume(itemID ItemID, count uint16) bool {
+func (t *slotTable) consume(itemID ItemID, count uint32) bool {
 	return t.consumeWithin(itemID, count, len(t))
 }
 
@@ -288,14 +288,24 @@ func (t *slotTable) consume(itemID ItemID, count uint16) bool {
 // on the player's legs because the pack ran out. The same bound answers "how much silver
 // is in the purse", for the plainer reason that a purse is a pack slot and a breastplate
 // is not.
-func (t *slotTable) consumePack(itemID ItemID, count uint16) bool {
+func (t *slotTable) consumePack(itemID ItemID, count uint32) bool {
 	return t.consumeWithin(itemID, count, equipmentFirst)
 }
 
 // consumeWithin is the spending rule itself, over the leading `limit` slots. One
 // implementation, for the reason `insert` is one: two loops that walk a slot table
 // subtracting counts are two places for "the stack emptied" to be got wrong.
-func (t *slotTable) consumeWithin(itemID ItemID, count uint16, limit int) bool {
+//
+// **The count is a uint32 because a spend is a pack-wide quantity, not a slot-wide
+// one.** A slot holds a uint16, but a spend walks many slots, so what is affordable is
+// bounded by the sum and not by any one stack — which is exactly why [slotTable.heldInPack]
+// reports a uint32. Taking a uint16 here made every caller with a wider total narrow it
+// first, and a narrowing on the paying side of a trade is a purchase settled for
+// `total mod 65536`.
+//
+// The narrowing that remains is the safe direction: `spent` is at most `stack.count`,
+// which is a uint16 by construction, so it fits one whatever `count` was.
+func (t *slotTable) consumeWithin(itemID ItemID, count uint32, limit int) bool {
 	if itemID == ItemNone || count == 0 {
 		return false
 	}
@@ -308,8 +318,8 @@ func (t *slotTable) consumeWithin(itemID ItemID, count uint16, limit int) bool {
 		if stack.item != itemID || stack.count == 0 {
 			continue
 		}
-		spent := min(count, stack.count)
-		stack.count -= spent
+		spent := min(count, uint32(stack.count))
+		stack.count -= uint16(spent)
 		count -= spent
 		if stack.count == 0 {
 			*stack = inventoryStack{}
